fixes: separate tests on several suits
